# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "这是一个动画网站的后端API服务",
        "title": "动画网站 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/carousels": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "创建新的轮播图记录",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "carousel"
                ],
                "summary": "创建新的轮播图",
                "parameters": [
                    {
                        "type": "string",
                        "description": "轮播图标题",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "轮播图副标题",
                        "name": "subtitle",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "轮播图详细描述",
                        "name": "description",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "轮播图图片文件",
                        "name": "image_file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "轮播图点击跳转链接",
                        "name": "link",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "轮播图显示顺序",
                        "name": "order",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "description": "轮播图是否激活",
                        "name": "is_active",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "轮播图开始显示时间",
                        "name": "start_date",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "轮播图结束显示时间",
                        "name": "end_date",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.CarouselResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/carousels/order": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "批量更新轮播图的显示顺序",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "carousel"
                ],
                "summary": "更新轮播图顺序",
                "parameters": [
                    {
                        "description": "轮播图顺序数组",
                        "name": "orders",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/controllers.CarouselOrderRequest"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/carousels/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "根据ID获取单个轮播图详情",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "carousel"
                ],
                "summary": "获取单个轮播图",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "轮播图ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CarouselResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "更新指定ID的轮播图信息",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "carousel"
                ],
                "summary": "更新轮播图",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "轮播图ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "轮播图标题",
                        "name": "title",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "轮播图副标题",
                        "name": "subtitle",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "轮播图详细描述",
                        "name": "description",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "轮播图图片文件",
                        "name": "image_file",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "轮播图点击跳转链接",
                        "name": "link",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "轮播图显示顺序",
                        "name": "order",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "description": "轮播图是否激活",
                        "name": "is_active",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "轮播图开始显示时间",
                        "name": "start_date",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "轮播图结束显示时间",
                        "name": "end_date",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CarouselResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "删除指定ID的轮播图",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "carousel"
                ],
                "summary": "删除轮播图",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "轮播图ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.SuccessResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/carousels": {
            "get": {
                "description": "获取所有轮播图列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "carousel"
                ],
                "summary": "获取所有轮播图",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.CarouselResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.CarouselOrderRequest": {
            "description": "更新轮播图顺序的请求参数",
            "type": "object",
            "required": [
                "id",
                "order"
            ],
            "properties": {
                "id": {
                    "description": "@Description 轮播图ID",
                    "type": "integer"
                },
                "order": {
                    "description": "@Description 轮播图顺序",
                    "type": "integer"
                }
            }
        },
...
        "controllers.ErrorResponse": {
            "description": "错误响应",
            "type": "object",
            "properties": {
                "error": {
                    "description": "@Description 错误信息",
                    "type": "string"
                }
            }
        },
...
        "controllers.SuccessResponse": {
            "description": "成功响应",
            "type": "object",
            "properties": {
                "message": {
                    "description": "@Description 成功信息",
                    "type": "string"
                }
            }
        },
...
        "models.CarouselResponse": {
            "description": "轮播图响应数据",
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "@Description 创建时间",
                    "type": "string"
                },
                "description": {
                    "description": "@Description 轮播图详细描述",
                    "type": "string"
                },
                "end_date": {
                    "description": "@Description 轮播图结束显示时间",
                    "type": "string"
                },
                "id": {
                    "description": "@Description 轮播图ID",
                    "type": "integer"
                },
                "image_url": {
                    "description": "@Description 轮播图图片URL",
                    "type": "string"
                },
                "is_active": {
                    "description": "@Description 轮播图是否激活",
                    "type": "boolean"
                },
                "link": {
                    "description": "@Description 轮播图点击跳转链接",
                    "type": "string"
                },
                "order": {
                    "description": "@Description 轮播图显示顺序",
                    "type": "integer"
                },
                "start_date": {
                    "description": "@Description 轮播图开始显示时间",
                    "type": "string"
                },
                "subtitle": {
                    "description": "@Description 轮播图副标题",
                    "type": "string"
                },
                "title": {
                    "description": "@Description 轮播图标题",
                    "type": "string"
                },
                "updated_at": {
                    "description": "@Description 更新时间",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "请在此输入 Bearer token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}